Extract writer link interactions
